+++ selene/_managed.py
@@ -1,6 +1,5 @@
 from selene.core.configuration import Config
-from selene.core.entity import Browser
-
+from selene.core._browser import Browser
 
 config = Config()
 browser = Browser(config)